requirements.txt
langchain>=0.1.16
groq
tiktoken
numpy
scikit-learn
sentence-transformers
streamlit
python-dotenv